# store-contract/src/main/resources/contract/openapi.yaml
openapi: 3.1.0
info:
  title: FengYu-WebSite API
  version: 1.0.0
  summary: Catalog, publishing, review and delivery API for the Infinia / FengYu ecosystem.
  description: |
    Versioned REST API of FengYu-WebSite (design §10).

    * Root path `/api/v1`; breaking changes ship as `/api/v2` — never via headers.
    * JSON is camelCase, timestamps are UTC RFC 3339, identifiers are UUIDv7/ULID.
    * Errors use `application/problem+json` with a stable `code`, `traceId` and
      localizable arguments.
    * Lists use cursor pagination; write endpoints accept `Idempotency-Key`;
      concurrent updates use `ETag/If-Match`.
  license:
    name: GPL-3.0
    url: https://www.gnu.org/licenses/gpl-3.0.txt
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /publisher/listings:
    get:
      operationId: listPublisherListings
      tags: [Publisher]
      summary: Listings owned by the caller's namespaces
      responses:
        '200':
          description: Listings
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/CatalogItem' }
    post:
      operationId: createListing
      tags: [Publisher]
      summary: Create a listing in a namespace the caller owns
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateListingRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ListingDetail' }
        '403': { $ref: '#/components/responses/ProblemForbidden' }
        '409': { $ref: '#/components/responses/ProblemConflict' }
...
  /publisher/releases/{releaseId}/uploads:
    post:
      operationId: createUploadSession
      tags: [Publisher]
      summary: Create an upload session; returns a short-lived presigned PUT URL
      parameters:
        - { name: releaseId, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [filename, kind]
              properties:
                filename: { type: string, maxLength: 255 }
                kind: { $ref: '#/components/schemas/ArtifactKind' }
                platform: { $ref: '#/components/schemas/Platform' }
                arch: { $ref: '#/components/schemas/Arch' }
                variant: { type: string, pattern: '^[a-z0-9][a-z0-9-]{0,31}$' }
                size: { type: integer, format: int64, minimum: 1 }
      responses:
        '201':
          description: Upload session
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UploadSession' }
...
  /publisher/releases/{releaseId}:
    get:
      operationId: getPublisherRelease
      tags: [Publisher]
      summary: Release status including scan findings
      parameters:
        - { name: releaseId, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: Release
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PublisherRelease' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    ProblemForbidden:
      description: Forbidden
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    ProblemConflict:
      description: Conflict
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    ListingType:
      type: string
      enum: [APP, PLUGIN, SKILL, MCP, FLOW]
    Channel:
      type: string
      enum: [stable, rc, beta, alpha, nightly, private]
    Platform:
      type: string
      enum: [windows, macos, linux, universal]
    Arch:
      type: string
      enum: [x64, arm64, universal]
    ArtifactKind:
      type: string
      enum: [PACKAGE, INSTALLER, PORTABLE, CHECKSUMS, SIGNATURE, SBOM]
    Problem:
      type: object
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        code: { type: string, description: Stable machine-readable code }
        detail: { type: string }
        traceId: { type: string }
        parameters:
          type: object
          additionalProperties: true
    CatalogItem:
      type: object
      required: [coordinate, type, namespace, slug, name]
      properties:
        coordinate: { type: string, example: 'infinia://plugin/official/markdown' }
        type: { $ref: '#/components/schemas/ListingType' }
        namespace: { type: string }
        slug: { type: string }
        name: { type: string }
        summary: { type: string }
        category: { type: string }
        tags: { type: array, items: { type: string } }
        iconUrl: { type: string, format: uri }
        latestVersion: { type: string }
        channel: { $ref: '#/components/schemas/Channel' }
        downloads: { type: integer, format: int64 }
        publisherName: { type: string }
        updatedAt: { type: string, format: date-time }
        featured: { type: boolean, description: Editorial featuring (admin) }
        minBeeLevel: { type: integer, minimum: 0, maximum: 4, description: 'Infinia Level gate: 0 = public; 1..4 restricts view/download to accounts at or above that level' }
...
    Localization:
      type: object
      properties:
        locale: { type: string, example: en }
        name: { type: string }
        summary: { type: string }
        descriptionMarkdown: { type: string }
        changelogMarkdown: { type: string }
    Artifact:
      type: object
      required: [artifactId, kind, platform, arch, filename, size]
      properties:
        artifactId: { type: string }
        kind: { $ref: '#/components/schemas/ArtifactKind' }
        platform: { $ref: '#/components/schemas/Platform' }
        arch: { $ref: '#/components/schemas/Arch' }
        variant: { type: string, description: APP build variant (lite, jre, uos, jar); default for non-APP artifacts }
        filename: { type: string }
        size: { type: integer, format: int64 }
        sha256: { type: string, nullable: true, description: Absent for live upstream delivery }
        keyId: { type: string }
        mimeType: { type: string }
    Dependency:
      type: object
      properties:
        coordinate: { type: string }
        range: { type: string }
        optional: { type: boolean }
    Permission:
      type: object
      properties:
        permissionId: { type: string }
        scope: { type: string, example: network }
        required: { type: boolean }
        reason: { type: string }
    ListingRelease:
      type: object
      required: [releaseId, version, status, channel]
      properties:
        releaseId: { type: string }
        version: { type: string }
        status: { type: string }
        channel: { $ref: '#/components/schemas/Channel' }
        publishedAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
        requiresHost: { type: string }
        license: { type: string }
        sourceUrl: { type: string }
        changelogMarkdown: { type: string }
        rolloutPercent: { type: integer, minimum: 0, maximum: 100 }
        artifacts: { type: array, items: { $ref: '#/components/schemas/Artifact' } }
        dependencies: { type: array, items: { $ref: '#/components/schemas/Dependency' } }
        permissions: { type: array, items: { $ref: '#/components/schemas/Permission' } }
    ListingDetail:
      type: object
      required: [listingId, coordinate, type, namespace, slug]
      properties:
        listingId: { type: string }
        coordinate: { type: string }
        type: { $ref: '#/components/schemas/ListingType' }
        namespace: { type: string }
        slug: { type: string }
        visibility: { type: string, enum: [PUBLIC, UNLISTED, PRIVATE] }
        status: { type: string }
        category: { type: string }
        tags: { type: array, items: { type: string } }
        iconUrl: { type: string, format: uri, nullable: true }
        screenshots: { type: array, items: { type: string, format: uri } }
        defaultChannel: { $ref: '#/components/schemas/Channel' }
        publisherName: { type: string }
        downloads: { type: integer, format: int64 }
        favorites: { type: integer, format: int64 }
        minBeeLevel: { type: integer, minimum: 0, maximum: 4, description: 'Infinia Level gate: 0 = public; 1..4 restricts view/download to accounts at or above that level' }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        localizations: { type: array, items: { $ref: '#/components/schemas/Localization' } }
        releases: { type: array, items: { $ref: '#/components/schemas/ListingRelease' } }
        upstream: { $ref: '#/components/schemas/UpstreamProvenance' }
    UpstreamProvenance:
      type: object
      description: Metadata provenance for an aggregated upstream artifact. The sync
        materializes the payload into the store's blob storage, so the entry ships
        like any publisher upload with a real, platform-signed digest.
      properties:
        sourceName: { type: string }
        externalId: { type: string }
        sourceUrl: { type: string, format: uri }
        sourcePath: { type: string }
        ref: { type: string }
        commitSha: { type: string }
        upstreamVersion: { type: string }
        metadataSha256: { type: string }
        firstSeenAt: { type: string, format: date-time }
        lastSeenAt: { type: string, format: date-time }
        deliveryMode: { type: string, enum: [MATERIALIZED_BLOB] }
...
    CreateListingRequest:
      type: object
      required: [namespace, slug, type, name]
      properties:
        namespace: { type: string, pattern: '^[a-z0-9][a-z0-9-]{0,62}$' }
        slug: { type: string, pattern: '^[a-z0-9][a-z0-9-]{0,62}$' }
        type: { $ref: '#/components/schemas/ListingType' }
        category: { type: string }
        tags: { type: array, items: { type: string } }
        defaultChannel: { $ref: '#/components/schemas/Channel' }
        name: { type: string, minLength: 1, maxLength: 100 }
        summary: { type: string, maxLength: 500 }
        descriptionMarkdown: { type: string, maxLength: 50000 }
        locale: { type: string, default: en }
        minBeeLevel: { type: integer, minimum: 0, maximum: 4, default: 0, description: 'Infinia Level gate the publisher sets: 0 = public; 1..4 = only signed-in accounts at or above that level may view and download' }
...
    UploadSession:
      type: object
      required: [uploadId, releaseId, uploadUrl, method]
      properties:
        uploadId: { type: string }
        releaseId: { type: string }
        filename: { type: string }
        kind: { $ref: '#/components/schemas/ArtifactKind' }
        platform: { $ref: '#/components/schemas/Platform' }
        arch: { $ref: '#/components/schemas/Arch' }
        variant: { type: string }
        uploadUrl: { type: string, format: uri }
        method: { type: string, enum: [PUT] }
        expiresAt: { type: string, format: date-time }
        maxSizeBytes: { type: integer, format: int64 }
    ScanFinding:
      type: object
      properties:
        severity: { type: string, enum: [INFO, WARN, ERROR, CRITICAL] }
        rule: { type: string }
        message: { type: string }
...
    PublisherRelease:
      type: object
      required: [releaseId, listingCoordinate, version, status, channel]
      properties:
        releaseId: { type: string }
        listingCoordinate: { type: string }
        version: { type: string }
        status: { type: string }
        channel: { $ref: '#/components/schemas/Channel' }
        createdAt: { type: string, format: date-time }
        publishedAt: { type: string, format: date-time, nullable: true }
        artifacts: { type: array, items: { $ref: '#/components/schemas/Artifact' } }
        findings: { type: array, items: { $ref: '#/components/schemas/ScanFinding' } }
...